// src/api/generated/model/getPeriodNutrientIntakeParams.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Fitness API
 * REST API for the Fitness & Nutrition SPA. Call GET /sanctum/csrf-cookie then POST /api/auth/login before accessing protected endpoints.
 * OpenAPI spec version: 1.0.0
 */
import type { GetPeriodNutrientIntakePeriod } from './getPeriodNutrientIntakePeriod';

export type GetPeriodNutrientIntakeParams = {
period?: GetPeriodNutrientIntakePeriod;
start_date?: string;
end_date?: string;
};